@@ -14,7 +14,7 @@ class UsersController < ApplicationController
   end
   
   def create
-    @user = User.new(params[:user])
+    @user = User.new(user_params)
     @user.admin = false
     
     if @user.save
@@ -30,7 +30,7 @@ class UsersController < ApplicationController
   
   def update
     @user.admin = params[:user][:admin] if admin?
-    if @user.update_attributes(params[:user])
+    if @user.update_attributes(user_params)
       flash[:notice] = "User successfully updated."
       redirect_to :action => "show", :id => @user.id
     else
@@ -91,5 +91,9 @@ class UsersController < ApplicationController
   
   def find_user
     @user = User.find(params[:id])
+  end
+
+  def user_params
+    params.require(:user).permit(:email, :name)
   end
 end
