@@ -1,7 +1,12 @@
 # Be sure to restart your server when you modify this file.
 
-# Your secret key for verifying the integrity of signed cookies.
+# Your secret key is used for verifying the integrity of signed cookies.
 # If you change this key, all old signed cookies will become invalid!
+
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Quotes::Application.config.secret_token = 'ed44615d50f51c550ece6280a8f43616dfab4b628310ea45fafc7c6226062ea5ac98d073cb36d26535a04291e68e2eea435ce834536315a9e99c796e4034a00b'
+# You can use `rake secret` to generate a secure secret key.
+
+# Make sure your secret_key_base is kept private
+# if you're sharing your code publicly.
+Quotes::Application.config.secret_key_base = '8a2357e10d639d2733eb7037a0290f0de1589b1980080683b359bf7d3686abded29667212437dfaf523c4edc9c833cfe0e41ee308461b29422fd897023d5ada5'
