@@ -2,10 +2,10 @@ class CreateProfiles < ActiveRecord::Migration
   def self.up
     create_table :profiles do |t|
       t.string :name, :default => '', :null => false
-      t.text :html, :default => '', :null => false
+      t.text :html, :null => false
       t.string :from, :default => '', :null => false
       t.string :subject, :default => '', :null => false
-      t.text :intro, :default => '', :null => false
+      t.text :intro, :null => false
 
       t.timestamps
     end
