@@ -2,7 +2,7 @@ class CreateItems < ActiveRecord::Migration
   def self.up
     create_table :items do |t|
       t.string :name, :default => '', :null => false
-      t.text :information, :default => '', :null => false
+      t.text :information, :null => false
       t.integer :price, :default => 0, :null => false
       t.integer :position, :default => 0, :null => false
 
